long init

+++ Java/library/structure/lazysegmenttree/RAMN.java
@@ -7,6 +7,6 @@ import static java.lang.Math.*;
  * 区間加算, 区間最小値
  */
 public final class RAMN extends LazySegmentTree {
-	public RAMN(final int[] a, final int e){ super(a, (x, y) -> min(x, y), (x, y) -> x + y, (x, y) -> x + y, e, 0); }
+	public RAMN(final int[] a, final long e){ super(a, (x, y) -> min(x, y), (x, y) -> x + y, (x, y) -> x + y, e, 0); }
 	public RAMN(final long[] a, final long e){ super(a, (x, y) -> min(x, y), (x, y) -> x + y, (x, y) -> x + y, e, 0); }
 }
