+++ Java/library/structure/lazysegmenttree/RUSM.java
@@ -9,7 +9,7 @@ import library.structure.pair.IntPair;
 public final class RUSM extends LazySegmentTreePair {
 	private final int n;
 	private final IntPair[] b;
-	public RUSM(final int[] a, final int id) {
+	public RUSM(final int[] a, final long id) {
 		super(a.length, (x, y) -> x.mul(y), (x, y) -> IntPair.of(x.first.longValue() + x.second.longValue() * y, x.second.longValue()), (x, y) -> x + y, IntPair.of(0, 0), id);
 		n = a.length;
 		b = new IntPair[n];
